requirements.txt
asgiref==3.7.2
beautifulsoup4==4.10.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
configobj==5.0.6
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.6.0
httplib2==0.20.2
idna==2.5
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.1.3
MarkupSafe==2.0.1
mock==4.0.3
more-itertools==8.10.0
mysqlclient==2.2.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==21.3
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
# pyparted==3.11.7
PyQt5==5.15.6
python-dotenv==0.19.1
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
# python-apt==2.4.0+ubuntu2
# python-debian==0.1.43+ubuntu1.1
# python-gnupg==0.4.8
python-magic==0.4.24
pytz==2022.1
pyxdg==0.27
QtPy==2.0.0
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.4
# systemd-python==234
tldextract==3.1.2
typing_extensions==4.8.0
# ubuntu-drivers-common==0.0.0
# ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
wadllib==1.3.6
# xdg==5
# xkit==0.0.0
# youtube-dl==2021.12.17
zipp==1.0.0
